big changes
dosya yapısı login register sayfaları ayrı ayrı oluşturuldu register içerisindeki yönlendirmeler ve kontorller yazıldı. karşılama sayfası oluşturuldu. logo ve resim eklendi ui devam edecek
