+++ day-story-web/package.json
@@ -8,7 +8,9 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.7.0",
+    "date-fns": "^3.6.0",
     "react": "^18.3.1",
+    "react-datepicker": "^6.9.0",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.23.1",
     "react-scripts": "5.0.1",
