+++ day-story-web/src/Pages/LoginPage/scss/loginPage.css
@@ -1,25 +1,31 @@
 @import url("https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap");
 .container {
-  padding: 60px;
+  padding: 43px 78px;
 }
 
 .header {
   display: flex;
   align-items: center;
-  gap: 10px;
 }
 .header__logo {
-  background-color: #79767D;
-  border-radius: 50%;
-  width: 60px;
-  height: 60px;
+  background-color: white;
+  width: 70px;
+  height: 70px;
 }
 .header__text {
   font-family: "Roboto", sans-serif;
+  color: rgb(90, 108, 165);
   font-weight: 400;
   font-size: 32px;
   line-height: 40px;
   margin: 0px;
+}
+.header__text span {
+  font-family: "Roboto", sans-serif;
+  font-size: 32px;
+  font-weight: 300;
+  line-height: 40px;
+  text-align: left;
 }
 
 .form {
@@ -27,19 +33,17 @@
   justify-content: space-around;
 }
 .form__description {
+  width: 579px;
   justify-content: center;
   align-items: center;
   text-align: center;
-  padding: 103px;
+  padding: 60px;
 }
 .form__description-img {
-  width: 400px;
-  height: 400px;
-  background-color: rgba(20, 18, 24, 0.08);
-  border-radius: 25px;
+  height: 482px;
 }
 .form__description-text {
-  width: 400px;
+  width: 579px;
 }
 .form__description-text h2 {
   color: rgb(0, 0, 0);
@@ -94,12 +98,15 @@
   align-items: center;
   text-align: center;
   display: block;
-  margin-top: 40px;
+  margin-top: 80px;
   width: 500px;
+  display: grid;
+  gap: 10px;
 }
 .form__list-footer md-filled-button {
   width: 500px;
 }
 .form__list-footer p {
+  margin-top: 40px;
   width: auto;
-}/*# sourceMappingURL=ReceptionPage.css.map */
+}/*# sourceMappingURL=loginPage.css.map */
