+++ day-story-web/src/Pages/RegisterPage/RegisterPage.jsx
@@ -1,0 +1,170 @@
+import '@material/web/all';
+import React, { useState } from 'react';
+import { Link, useNavigate } from "react-router-dom";
+
+import './scss/RegisterPage.css';
+
+//components
+import FormListFirst from './formlistfirst';
+import FormListSecond from './formlistsecond';
+
+//images
+import logo from '../../../src/assets/images/daystory-logo.png';
+import günlük from '../../../src/assets/images/daystory.png';
+
+
+const RegisterPage = () => {
+  const navigate = useNavigate();
+  const [loading, setLoading] = useState(false);
+  const [currentForm, setCurrentForm] = useState(1);
+
+  const [formData, setFormData] = useState({
+    firstName: "",
+    lastName: "",
+    gender: "",
+    birthdate: "",
+    email: "",
+    username: "",
+    password: "",
+    role: ["default"],
+  });
+
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setFormData((prevData) => ({
+      ...prevData,
+      [name]: value
+    }));
+  };
+
+  const handleGenderChange = (e) => {
+    const { value } = e.target;
+    setFormData((prevData) => ({
+      ...prevData,
+      gender: value
+    }));
+  };
+
+  const validateForm1 = () => {
+    if (!formData.firstName.trim()) {
+      alert("İsim boş bırakılamaz.");
+      return false;
+    }
+    if (!formData.lastName.trim()) {
+      alert("Soyisim boş bırakılamaz.");
+      return false;
+    }
+    if (!formData.gender.trim()) {
+      alert("Cinsiyet boş bırakılamaz.");
+      return false;
+    }
+    if (!formData.birthdate.trim()) {
+      alert("Doğum Tarihi boş bırakılamaz.");
+      return false;
+    }
+    return true;
+  };
+
+  const validateForm2 = () => {
+    if (!formData.email.trim()) {
+      alert("Email boş bırakılamaz.");
+      return false;
+    }
+    if (!formData.username.trim()) {
+      alert("Kullanıcı adı boş bırakılamaz.");
+      return false;
+    }
+    if (!formData.password.trim()) {
+      alert("Şifre boş bırakılamaz.");
+      return false;
+    }
+    return true;
+  };
+
+  const registerUser = async () => {
+    if (!validateForm2()) return;
+
+    try {
+      const response = await fetch("", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json"
+        },
+        body: JSON.stringify(formData)
+      });
+
+      if (response.ok) {
+        setLoading(true);
+        navigate("/girişyap");
+      } else {
+        const errorData = await response.json();
+        console.error("Registration failed:", errorData);
+      }
+    } catch (error) {
+      console.error("Error registering user:", error);
+    }
+  };
+
+  const toggleForm = () => {
+    if (currentForm === 1) {
+      if (validateForm1()) {
+        setCurrentForm(2);
+      }
+    } else {
+      setCurrentForm(1);
+    }
+  };
+
+
+  return (
+    <div className='container'>
+      <header className='header'>
+        <img className='header__logo' src={logo} alt="daystory-logo" />
+        <p className='header__text'>Day<span>Story</span></p>
+      </header>
+
+      <div className='form'>
+        <div className='form__description'>
+          <img className='form__description-img' src={günlük} alt="main_image" />
+          <div className='form__description-text'>
+            <h2>Her sayfanın bir hikaye anlattığı yer</h2>
+            <p>Sizin anılarınızı kalıcı hale getiriyoruz</p>
+          </div>
+        </div>
+
+        <div className='form__list'>
+
+          {currentForm === 1 && <FormListFirst formData={formData} handleChange={handleChange} handleGenderChange={handleGenderChange} />}
+          {currentForm === 2 && <FormListSecond formData={formData} handleChange={handleChange} registerUser={registerUser} loading={loading} />}
+          
+          <div className='form__list-footer'>
+            <md-filled-button type="button" onClick={toggleForm} >
+              {currentForm === 1 ? 'Devam' : 'Geri'}
+            </md-filled-button>
+
+            {currentForm === 2 && (
+              <md-filled-button type="button" onClick={registerUser} >
+                Kayıt ol
+              </md-filled-button>
+            )}
+
+            <p>Zaten bir hesabın var mı?<Link to="#" >Giriş Yap</Link> </p>
+
+
+          </div>
+
+          <p>{formData.firstName}</p>
+          <p>{formData.lastName}</p>
+          <p>{formData.gender}</p>
+          <p>{formData.birthdate}</p>
+          <p>{formData.email}</p>
+          <p>{formData.username}</p>
+          <p>{formData.password}</p>
+
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default RegisterPage;
